% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_transformations_mult.R
\name{.computeLengthOfRemainingStick}
\alias{.computeLengthOfRemainingStick}
\title{Computes Length Of Remaining Stick}
\usage{
.computeLengthOfRemainingStick(theta_mat, k, hyp_direction)
}
\arguments{
\item{theta_mat}{matrix with samples from truncated Dirichlet density}

\item{k}{current parameter index}

\item{hyp_direction}{specifies whether the imposed inequality constrained imposes an increasing (i.e., 'smaller') or
decreasing (i.e., 'larger') trend}
}
\value{
stick length
}
\description{
When applying the probit transformation on the Dirichlet samples, this function is used as part of the stick-breaking
algorithm. It computes the length of the remaining stick when the current element is broken off.
}
