% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmult_transformations_binom.R
\name{tbinom_trans}
\alias{tbinom_trans}
\title{Transforms Truncated Beta Samples To Real Line}
\usage{
tbinom_trans(theta_mat, boundaries, binom_equal, hyp_direction)
}
\arguments{
\item{theta_mat}{matrix with samples from truncated beta density}

\item{boundaries}{list containing indices for upper and lower truncation boundaries}

\item{binom_equal}{multiplicative elements for each lower and upper bound of each inequality constrained parameter.}

\item{hyp_direction}{specifies whether the imposed inequality constrained imposes an increasing (i.e., 'smaller') or
decreasing (i.e., 'larger') trend}
}
\value{
matrix with transformed samples
}
\description{
Transforms samples from a truncated beta density to the real line using a stick-breaking algorithm.
This algorithm is suitable for mixtures of equality constrained parameters, inequality constrained parameters,
and free parameters
}
